Added rotate plugin
It allows to rotate an image, according to any angle.
To do this, you  must call function caman.rotate(degrees), where
degrees must be a number (positive or negative, on depends if you
want that your image rotates in clockwise direction or
counterclockwise direction).
If degrees is a multiple of 90 canvas mantains the dimensions
of image, otherwise the canvas will become larger, with height and
width equals to the diagonal of the image.